[Feature][seatunnel-translation] Add spark-common module to implement support spark multiple version
