+++ seatunnel-core/seatunnel-core-spark/pom.xml
@@ -68,7 +68,13 @@
 
         <dependency>
             <groupId>org.apache.seatunnel</groupId>
-            <artifactId>seatunnel-translation-spark</artifactId>
+            <artifactId>seatunnel-translation-spark-common</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+
+        <dependency>
+            <groupId>org.apache.seatunnel</groupId>
+            <artifactId>seatunnel-translation-spark-2.4</artifactId>
             <version>${project.version}</version>
         </dependency>
