+++ seatunnel-core/seatunnel-spark-starter/pom.xml
@@ -68,7 +68,7 @@
 
         <dependency>
             <groupId>org.apache.seatunnel</groupId>
-            <artifactId>seatunnel-translation-spark</artifactId>
+            <artifactId>seatunnel-translation-spark-2.4</artifactId>
             <version>${project.version}</version>
         </dependency>
