+++ seatunnel-translation/seatunnel-translation-spark/seatunnel-translation-spark-2.4/src/main/resources/META-INF/services/org.apache.spark.sql.sources.DataSourceRegister
